refactor(AWS Deploy): Generalize `uploadZipFile` logic
