+++ lib/plugins/aws/deploy/index.js
@@ -10,6 +10,7 @@ const checkIfBucketExists = require('../lib/check-if-bucket-exists');
 const getCreateChangeSetParams = require('../lib/get-create-change-set-params');
 const getExecuteChangeSetParams = require('../lib/get-execute-change-set-params');
 const waitForChangeSetCreation = require('../lib/wait-for-change-set-creation');
+const uploadZipFile = require('../lib/upload-zip-file');
 const createStack = require('./lib/create-stack');
 const cleanupS3Bucket = require('./lib/cleanup-s3-bucket');
 const uploadArtifacts = require('./lib/upload-artifacts');
@@ -47,6 +48,7 @@ class AwsDeploy {
       monitorStack,
       checkIfBucketExists,
       waitForChangeSetCreation,
+      uploadZipFile,
       getCreateChangeSetParams,
       getExecuteChangeSetParams
     );
