+++ lib/plugins/aws/deploy/lib/upload-artifacts.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const _ = require('lodash');
-const fs = require('fs');
 const fsp = require('fs').promises;
 const path = require('path');
 const crypto = require('crypto');
@@ -10,6 +9,7 @@ const filesize = require('filesize');
 const normalizeFiles = require('../../lib/normalize-files');
 const getLambdaLayerArtifactPath = require('../../utils/get-lambda-layer-artifact-path');
 const ServerlessError = require('../../../../serverless-error');
+const setS3UploadEncryptionOptions = require('../../../../aws/set-s3-upload-encryption-options');
 const { progress, log } = require('@serverless/utils/log');
 
 const MAX_CONCURRENT_ARTIFACTS_UPLOADS =
@@ -69,7 +69,7 @@ module.exports = {
 
     const deploymentBucketObject = this.serverless.service.provider.deploymentBucketObject;
     if (deploymentBucketObject) {
-      params = setServersideEncryptionOptions(params, deploymentBucketObject);
+      params = setS3UploadEncryptionOptions(params, deploymentBucketObject);
     }
 
     return this.provider.request('S3', 'upload', params);
@@ -95,47 +95,10 @@ module.exports = {
 
     const deploymentBucketObject = this.serverless.service.provider.deploymentBucketObject;
     if (deploymentBucketObject) {
-      params = setServersideEncryptionOptions(params, deploymentBucketObject);
+      params = setS3UploadEncryptionOptions(params, deploymentBucketObject);
     }
 
     return this.provider.request('S3', 'upload', params);
-  },
-
-  async uploadZipFile({ filename, s3KeyDirname }) {
-    const fileName = filename.split(path.sep).pop();
-
-    // TODO refactor to be async (use util function to compute checksum async)
-    const data = fs.readFileSync(filename);
-    const fileHash = crypto.createHash('sha256').update(data).digest('base64');
-
-    const artifactStream = fs.createReadStream(filename);
-    // As AWS SDK request might be postponed (requests are queued)
-    // eventual stream error may crash the process (it's thrown as uncaught if not observed).
-    // Below lines prevent that
-    let streamError;
-    artifactStream.on('error', (error) => (streamError = error));
-
-    let params = {
-      Bucket: this.bucketName,
-      Key: `${s3KeyDirname}/${fileName}`,
-      Body: artifactStream,
-      ContentType: 'application/zip',
-      Metadata: {
-        filesha256: fileHash,
-      },
-    };
-
-    const deploymentBucketObject = this.serverless.service.provider.deploymentBucketObject;
-    if (deploymentBucketObject) {
-      params = setServersideEncryptionOptions(params, deploymentBucketObject);
-    }
-
-    const response = await this.provider.request('S3', 'upload', params);
-    // Interestingly, if request handling was queued, and stream errored (before being consumed by
-    // AWS SDK) then SDK call succeeds without actually uploading a file to S3 bucket.
-    // Below line ensures that eventual stream error is communicated
-    if (streamError) throw streamError;
-    return response;
   },
 
   async getFunctionArtifactFilePaths() {
@@ -245,23 +208,3 @@ module.exports = {
     }
   },
 };
-
-function setServersideEncryptionOptions(putParams, deploymentBucketOptions) {
-  const encryptionFields = [
-    ['serverSideEncryption', 'ServerSideEncryption'],
-    ['sseCustomerAlgorithim', 'SSECustomerAlgorithm'],
-    ['sseCustomerKey', 'SSECustomerKey'],
-    ['sseCustomerKeyMD5', 'SSECustomerKeyMD5'],
-    ['sseKMSKeyId', 'SSEKMSKeyId'],
-  ];
-
-  const params = putParams;
-
-  encryptionFields.forEach((element) => {
-    if (deploymentBucketOptions[element[0]]) {
-      params[element[1]] = deploymentBucketOptions[element[0]];
-    }
-  }, this);
-
-  return params;
-}
